Completed integration of unity analytics. Completed auto aim and damage for enemies/player
Completed integration of unity analytics. As well as auto aim for when using one joystick or gyro. Implemented the a simple health manager for player and have enemies destroyed when hit with a bullet. Have an analytics sent at the end of the game loop, with time survived and the amount of enemies killed.
Fix #29 Fix #28 Fix #6